Renamed stdfblib/events files.
