+++ src/stdfblib/events/ATimeOut_atp.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2014 fortiss GmbH
+ * Copyright (c) 2013 fortiss GmbH
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
  * http://www.eclipse.org/legal/epl-2.0.
@@ -10,25 +10,25 @@
  *   Alois Zoitl - initial API and implementation and/or initial documentation
  *******************************************************************************/
 
-#include "ARTimeOut_adp.h"
+#include "ATimeOut_atp.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "ARTimeOut_adp_gen.cpp"
+#include "ATimeOut_atp_gen.cpp"
 #endif
 
 #include "criticalregion.h"
 #include "resource.h"
 
-DEFINE_ADAPTER_TYPE(FORTE_ARTimeOut, g_nStringIdARTimeOut)
+DEFINE_ADAPTER_TYPE(FORTE_ATimeOut, g_nStringIdATimeOut)
 
-const CStringDictionary::TStringId FORTE_ARTimeOut::scmDataOutputNames[] = {g_nStringIdDT};
-const CStringDictionary::TStringId FORTE_ARTimeOut::scmDataOutputTypeIds[] = {g_nStringIdTIME};
-const TForteInt16 FORTE_ARTimeOut::scmEIWithIndexes[] = {-1};
-const CStringDictionary::TStringId FORTE_ARTimeOut::scmEventInputNames[] = {g_nStringIdTimeOut};
-const TDataIOID FORTE_ARTimeOut::scmEOWith[] = {0, scmWithListDelimiter};
-const TForteInt16 FORTE_ARTimeOut::scmEOWithIndexes[] = {0, -1};
-const CStringDictionary::TStringId FORTE_ARTimeOut::scmEventOutputNames[] = {g_nStringIdSTART, g_nStringIdSTOP};
+const CStringDictionary::TStringId FORTE_ATimeOut::scmDataOutputNames[] = {g_nStringIdDT};
+const CStringDictionary::TStringId FORTE_ATimeOut::scmDataOutputTypeIds[] = {g_nStringIdTIME};
+const TForteInt16 FORTE_ATimeOut::scmEIWithIndexes[] = {-1};
+const CStringDictionary::TStringId FORTE_ATimeOut::scmEventInputNames[] = {g_nStringIdTimeOut};
+const TDataIOID FORTE_ATimeOut::scmEOWith[] = {0, scmWithListDelimiter};
+const TForteInt16 FORTE_ATimeOut::scmEOWithIndexes[] = {0, -1};
+const CStringDictionary::TStringId FORTE_ATimeOut::scmEventOutputNames[] = {g_nStringIdSTART, g_nStringIdSTOP};
 
-const SFBInterfaceSpec FORTE_ARTimeOut::scmFBInterfaceSpecSocket = {
+const SFBInterfaceSpec FORTE_ATimeOut::scmFBInterfaceSpecSocket = {
   1, scmEventInputNames, nullptr, scmEIWithIndexes,
   2, scmEventOutputNames, scmEOWith, scmEOWithIndexes,
   0, nullptr, nullptr,
@@ -37,7 +37,7 @@ const SFBInterfaceSpec FORTE_ARTimeOut::scmFBInterfaceSpecSocket = {
   0, nullptr
 };
 
-const SFBInterfaceSpec FORTE_ARTimeOut::scmFBInterfaceSpecPlug = {
+const SFBInterfaceSpec FORTE_ATimeOut::scmFBInterfaceSpecPlug = {
   2, scmEventOutputNames, scmEOWith, scmEOWithIndexes,
   1, scmEventInputNames, nullptr, scmEIWithIndexes,
   1, scmDataOutputNames, scmDataOutputTypeIds,
@@ -46,7 +46,7 @@ const SFBInterfaceSpec FORTE_ARTimeOut::scmFBInterfaceSpecPlug = {
   0, nullptr
 };
 
-void FORTE_ARTimeOut::readInputData(const TEventID paEIID) {
+void FORTE_ATimeOut::readInputData(const TEventID paEIID) {
   if(isSocket()) {
     // nothing to do
   } else {
@@ -61,7 +61,7 @@ void FORTE_ARTimeOut::readInputData(const TEventID paEIID) {
   }
 }
 
-void FORTE_ARTimeOut::writeOutputData(const TEventID paEIID) {
+void FORTE_ATimeOut::writeOutputData(const TEventID paEIID) {
   if(isSocket()) {
     switch(paEIID) {
       case scmEventSTARTID: {
