+++ src/stdfblib/events/CMakeLists.txt
@@ -22,7 +22,7 @@ forte_add_sourcefile_hcpp(E_CTU E_CTD E_CTUD E_CTUD_UDINT_fbt E_DELAY  E_D_FF   
 forte_add_sourcefile_hcpp( E_REND   E_RS     E_SELECT  E_SR)
 forte_add_sourcefile_hcpp(E_CYCLE   E_DEMUX  E_F_TRIG  E_PERMIT) 
 forte_add_sourcefile_hcpp(E_RDELAY   E_RESTART E_R_TRIG E_SPLIT   E_SWITCH)
-forte_add_sourcefile_hcpp(GEN_E_DEMUX GEN_E_MUX)
-forte_add_sourcefile_hcpp(ATimeOut_adp E_TimeOut_fbt ARTimeOut_adp E_RTimeOut_fbt E_T_FF)
+forte_add_sourcefile_hcpp(GEN_E_DEMUX_fbt GEN_E_MUX_fbt)
+forte_add_sourcefile_hcpp(ATimeOut_atp E_TimeOut_fbt ARTimeOut_atp E_RTimeOut_fbt E_T_FF)
 forte_add_sourcefile_hcpp(E_TRAIN)
 forte_add_sourcefile_hcpp(E_PULSE_fbt E_TP_fbt E_TON_fbt E_TOF_fbt E_TONOF_fbt)
