+++ src/stdfblib/events/E_TimeOut_fbt.h
@@ -19,7 +19,7 @@
 #include "forte_array.h"
 #include "forte_array_fixed.h"
 #include "forte_array_variable.h"
-#include "ATimeOut_adp.h"
+#include "ATimeOut_atp.h"
 
 #include "../arch/timerha.h"
